[AV2-941] corrected all bad types
