+++ src/app/auth/auth-direct-login/auth-direct-login.component.spec.ts
@@ -119,7 +119,10 @@ describe('AuthDirectLoginComponent', () => {
       tick(50);
       fixture.detectChanges();
       expect(notificationSpy.alert.calls.count()).toBe(1);
-      notificationSpy.alert.calls.first().args[0].buttons[0].handler();
+
+      const button = notificationSpy.alert.calls.first().args[0].buttons[0];
+      (typeof button == 'string') ? button : button.handler(true);
+
       expect(routerSpy.navigate.calls.first().args[0]).toEqual(['login']);
     }));
