+++ src/app/auth/auth-logout/auth-logout.component.spec.ts
@@ -61,7 +61,7 @@ describe('AuthLogoutComponent', () => {
         fixture.detectChanges();
         component.onEnter();
         expect(newRelicSpy.setPageViewName).toHaveBeenCalledWith('logout');
-        authSpy.logout.and.returnValue({});
+        authSpy.logout.and.returnValue(Promise.resolve(true));
         expect(authSpy.logout.calls.count()).toBe(1);
     }));
 });
