+++ src/app/auth/auth-reset-password/auth-reset-password.component.spec.ts
@@ -83,7 +83,10 @@ describe('AuthResetPasswordComponent', () => {
       fixture.detectChanges();
       expect(notificationSpy.alert.calls.count()).toBe(1);
       expect(notificationSpy.alert.calls.first().args[0].message).toContain('Invalid');
-      notificationSpy.alert.calls.first().args[0].buttons[0].handler();
+
+      const button = notificationSpy.alert.calls.first().args[0].buttons[0];
+      (typeof button == 'string') ? button : button.handler(true);
+
       expect(routerSpy.navigate.calls.first().args[0]).toEqual(['login']);
     });
   });
