+++ src/testing/fixtures.ts
@@ -1,3 +1,3 @@
 export { SubmissionFixture } from './fixtures/assessment-submissions';
 export { ProgramFixture } from './fixtures/programs';
-export { mockMembers } from './fixtures/chats';
+export { mockMembers, mockChats } from './fixtures/chats';
