+++ src/testing/fixtures/chats.ts
@@ -1,6 +1,6 @@
-import { ChannelMembers } from "@app/chat/chat.service";
+import { ChannelMembers, ChatChannel } from "@app/chat/chat.service";
 
-export const mockMembers: ChannelMembers[] = [
+const mockMembers: ChannelMembers[] = [
   {
     uuid: '1',
     name: 'student+01',
@@ -23,3 +23,50 @@ export const mockMembers: ChannelMembers[] = [
     avatar: 'https://www.gravatar.com/avatar/21b7427270a606e8a3c4413a13bb47c6?d=https://sandbox.practera.com/img/user-512.png&s=50'
   }
 ];
+
+const channel1: ChatChannel = {
+  uuid: '35326928',
+  name: 'Team 1',
+  avatar: 'https://sandbox.practera.com/img/team-white.png',
+  pusherChannel: 'sdb746-93r7dc-5f44eb4f',
+  isAnnouncement: false,
+  isDirectMessage: false,
+  readonly: false,
+  roles: [
+    'participant',
+    'coordinator',
+    'admin'
+  ],
+  unreadMessageCount: 0,
+  lastMessage: null,
+  lastMessageCreated: null,
+  canEdit: true,
+};
+
+const channel2: ChatChannel = {
+  uuid: 'ced963c1',
+  name: 'Team 1 + Mentor',
+  avatar: 'https://sandbox.practera.com/img/team-white.png',
+  pusherChannel: 'kb5gt-9nfbj-5f45eb4g',
+  isAnnouncement: false,
+  isDirectMessage: false,
+  readonly: false,
+  roles: [
+    'participant',
+    'mentor',
+    'coordinator',
+    'admin'
+  ],
+  unreadMessageCount: 0,
+  lastMessage: null,
+  lastMessageCreated: null,
+  canEdit: true,
+};
+
+const mockChats = {
+  data: {
+    channels: [channel1, channel2]
+  }
+};
+
+export { mockChats, mockMembers };
